+++ baRSS/Constants.h
@@ -34,6 +34,8 @@
 static NSPasteboardType const UTI_OPML = @"org.opml";
 /// URL with newest baRSS releases. Automatically added when user starts baRSS for the first time.
 static NSString* const versionUpdateURL = @"https://github.com/relikd/baRSS/releases.atom";
+/// URL to help page of auxiliary application "URL Scheme Defaults"
+static NSString* const auxiliaryAppURL = @"https://github.com/relikd/URL-Scheme-Defaults#url-scheme-defaults";
 
 
 #pragma mark - NSImageName constants